Move schemas to session package
