+++ backend/t4cclient/routes/users.py
@@ -16,7 +16,7 @@ from t4cclient.schemas.repositories.users import (
     PatchUserRoleRequest,
     Role,
 )
-from t4cclient.schemas.sessions import AdvancedSessionResponse
+from t4cclient.sessions.schema import AdvancedSessionResponse
 
 router = APIRouter()
 
@@ -68,7 +68,7 @@ def update_role_of_user(
         repository_users.delete_all_repositories_for_user(db, username)
     return users.update_role_of_user(db, username, body.role)
 
-
+# TODO: This is actually a sessions route (sessions/{username}?)
 @router.get("/{username}/sessions", response_model=t.List[AdvancedSessionResponse])
 def get_sessions_for_user(
     username: str, db: Session = Depends(get_db), token=Depends(JWTBearer())
