+++ backend/t4cclient/sessions/sessions.py
@@ -13,7 +13,7 @@ import requests
 # local:
 from t4cclient import config
 from t4cclient.core.operators import OPERATOR
-from t4cclient.schemas.sessions import WorkspaceType
+from t4cclient.sessions.schema import WorkspaceType
 from t4cclient.sessions.models import DatabaseSession
 
 log = logging.getLogger(__name__)
